repo_fixtures to fixtures renamed
