+++ backend/testhelpers/fixtures/join_request.go
@@ -1,4 +1,4 @@
-package repo_fixtures
+package fixtures
 
 import (
 	"fmt"
