+++ backend/testhelpers/fixtures/member.go
@@ -1,4 +1,4 @@
-package repo_fixtures
+package fixtures
 
 import (
 	"fmt"
